Rename interpolation_new
